Refactoring. Naming things.
