--- test_foobar.py
@@ -1,7 +1,0 @@
-from foobar import FooBarCommand
-
-
-def test_foobar():
-    foobar = FooBarCommand()
-    result = foobar("hello")
-    assert result == "hello: foobar"
